+++ .idea/compiler.xml
@@ -6,6 +6,7 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
+        <module name="JPA-spring" />
         <module name="JPA-Spring" />
       </profile>
     </annotationProcessing>
