--- .idea/libraries/h2_1_4_200.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="h2-1.4.200">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/h2/bin/h2-1.4.200.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
