+++ src/main/java/helloJpa/entity/Member.java
@@ -3,6 +3,7 @@ package helloJpa.entity;
 import javax.persistence.*;
 import java.util.Date;
 
+@Entity
 public class Member {
 
     @Id //primary key
