+++ src/main/java/helloJpa/relationEntity/RelationMember.java
@@ -2,6 +2,7 @@ package helloJpa.relationEntity;
 
 import javax.persistence.*;
 
+@Entity
 public class RelationMember {
 
     @Id @GeneratedValue(strategy = GenerationType.AUTO)
