+++ src/main/java/helloJpa/relationEntity/RelationTeam.java
@@ -7,6 +7,7 @@ import javax.persistence.OneToMany;
 import java.util.ArrayList;
 import java.util.List;
 
+@Entity
 public class RelationTeam {
 
     @Id @GeneratedValue
