--- src/main/java/reJPA/entity/Member.java
@@ -1,88 +1,0 @@
-package reJPA.entity;
-
-import javax.persistence.*;
-import java.util.Date;
-
-@Entity
-public class Member {
-
-    @Id @GeneratedValue(strategy=GenerationType.AUTO)
-    private Long id;
-
-    @Column(name = "USERNAME")
-    private String name;
-    private int age;
-
-//    @Column(name = "TEAM_ID")
-//    private Long teamId;
-
-    @ManyToOne
-    @JoinColumn(name="TEAM_ID")
-    private Team team;
-
-    @Enumerated(EnumType.STRING)
-    private MemberType memberType;
-
-    @Temporal(TemporalType.TIMESTAMP)
-    private Date regDate;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    public MemberType getMemberType() {
-        return memberType;
-    }
-
-    public void setMemberType(MemberType memberType) {
-        this.memberType = memberType;
-    }
-
-    public Date getRegDate() {
-        return regDate;
-    }
-
-    public void setRegDate(Date regDate) {
-        this.regDate = regDate;
-    }
-
-    public Team getTeam() {
-        return team;
-    }
-
-    public void setTeam(Team team) {
-        this.team = team;
-    }
-
-    @Override
-    public String toString() {
-        return "Member{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", age=" + age +
-                ", team=" + team +
-                ", memberType=" + memberType +
-                ", regDate=" + regDate +
-                '}';
-    }
-}
