+++ src/main/resources/META-INF/persistence.xml
@@ -1,12 +1,18 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <persistence xmlns="http://xmlns.jcp.org/xml/ns/persistence" version="2.2">
-    <persistence-unit name="helloJPA">
+    <persistence-unit name="hello">
         <properties>
             <!-- 필수 속성 -->
             <property name="javax.persistence.jdbc.driver" value="org.h2.Driver"/>
             <property name="javax.persistence.jdbc.user" value="sa"/>
             <property name="javax.persistence.jdbc.password" value=""/>
+
+            <!--
             <property name="javax.persistence.jdbc.url" value="jdbc:h2:tcp://localhost/~/test"/>
+            -->
+
+            <property name="javax.persistence.jdbc.url" value="jdbc:h2:~/test"/>
+
             <!-- 방언 설정 -->
             <property name="hibernate.dialect" value="org.hibernate.dialect.H2Dialect"/>
             <!--
@@ -23,7 +29,7 @@
             <property name="hibernate.hbm2ddl.auto" value="create" />
             <!--
             create: 기존테이블 삭제 후 다시 생성 (DROP+CREATE)- 운영DB xx,
-            createㄱ-drop: create와 같으나 종료시점에 테이블 drop - 운영DB xx,
+            create-drop: create와 같으나 종료시점에 테이블 drop - 운영DB xx,
             update: 변경분만 반영 - 운영DB xx,
             validate: 엔티티와 테이블이 정상 매핑되었는지만 확인
             none: 사용하지 않음
